Separate api data classes, move to data package
