+++ db/src/main/scala/squeryl/ValidatorDbApiImpl.scala
@@ -4,7 +4,7 @@ import cats.effect.Sync
 import cats.syntax.all.*
 import sdk.CustomTypeMode.*
 import sdk.RhonixNodeDb.validatorTable
-import sdk.api.Validator
+import sdk.api.data.Validator
 import sdk.db.DbSession.withSessionF
 import sdk.db.{DbSession, ValidatorDbApi, ValidatorTable}
