+++ sdk/src/main/scala/sdk/api/ExternalApi.scala
@@ -1,5 +1,7 @@
 package sdk.api
 
+import sdk.api.data.*
+
 trait ExternalApi[F[_]] {
   def getBlockById(blockId: Long): F[Option[Block]]
   def getDeploysByBlockId(blockId: Long): F[Seq[Deploy]]
