+++ sdk/src/main/scala/sdk/api/data/BlockBonds.scala
@@ -1,0 +1,9 @@
+package sdk.api.data
+
+import cats.kernel.Eq
+
+final case class BlockBonds(blockId: Long, bondId: Long)
+
+object BlockBonds {
+  implicit val bondEq: Eq[BlockBonds] = Eq.fromUniversalEquals
+}
