+++ sdk/src/main/scala/sdk/api/data/BlockDeploys.scala
@@ -1,0 +1,3 @@
+package sdk.api.data
+
+final case class BlockDeploys(blockId: Long, deployId: Long)
