+++ sdk/src/main/scala/sdk/api/data/BlockJustifications.scala
@@ -1,0 +1,3 @@
+package sdk.api.data
+
+final case class BlockJustifications(validatorId: Long, latestBlockId: Long)
