+++ sdk/src/main/scala/sdk/api/data/Bond.scala
@@ -1,0 +1,3 @@
+package sdk.api.data
+
+final case class Bond(validator: Validator, stake: Long)
