+++ sdk/src/main/scala/sdk/api/data/Justification.scala
@@ -1,0 +1,7 @@
+package sdk.api.data
+
+final case class Justification(
+  id: Long,
+  validator: Validator,
+  lastestBlock: Block,
+)
