+++ sdk/src/main/scala/sdk/api/data/Validator.scala
@@ -1,0 +1,12 @@
+package sdk.api.data
+
+final case class Validator(
+  publicKey: Array[Byte], // Unique index
+  http: String,
+) {
+  override def equals(obj: Any): Boolean = obj match {
+    case that: Validator =>
+      this.publicKey.sameElements(that.publicKey) && this.http == that.http
+    case _               => false
+  }
+}
