+++ sdk/src/main/scala/sdk/db/BlockTable.scala
@@ -1,7 +1,7 @@
 package sdk.db
 
 import sdk.DbTable
-import sdk.api.{Block, Bond, Deploy, Validator}
+import sdk.api.data.*
 
 @SuppressWarnings(Array("org.wartremover.warts.FinalCaseClass"))
 case class BlockTable(
